Feat/dynamic types

+++ src/IR/irGen/statements/SwitchStatementGenerationStrategy/SwitchStatementGenerationStrategy.cpp
@@ -1,3 +1,22 @@
+/*
+ * FlowWing Compiler
+ * Copyright (C) 2023-2025 Kushagra Rathore
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program; if not, write to the Free Software Foundation, Inc.,
+ * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+
 #include "SwitchStatementGenerationStrategy.h"
 #include <cstdint>
 #include <memory>
@@ -129,7 +148,7 @@ llvm::Value *SwitchStatementGenerationStrategy::generateGlobalStatement(
 void SwitchStatementGenerationStrategy::populateSwitchStatementsBasicBlock(
     std::vector<llvm::BasicBlock *> &casesBB, const uint64_t &casesSize,
     llvm::Function *&function) {
-  for (int i = 1; i <= casesSize; i++) {
+  for (size_t i = 1; i <= casesSize; i++) {
     std::string caseName = "case:";
 
     if (i != casesSize) {
